Noob News
There are a huge collection of games out there, each game special and unique in some way. They all have something that makes it interesting for the player. As of right now there is no “one stop shop” for game information driven completely by players.
This is where Noob News comes into play. The online web application will allow users to search for their favorite games & be able to communicate with others through the game’s personal comment section. Each game will be given a rating which will be based on user feedback which will be used to generate a dynamic Top 40 list. Users are able to view popular let's play videos of the game (Youtube) as well as a live stream of the game if available (Twitch).
Any-one can go from Noob to Expert in no time.